refactor: Move initialisation inside WithABProvider
